+++ repldex/backend/static/markdownshortcuts.js
