+++ repldex/config/editors.txt
